cognitivehealthmonitoring/edgesim/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-cosmos>=4.0.0
requests>=2.25.0
